chore: Release v4.1.0
## Summary
- **Hybrid Embedding Strategy**: OpenAI online + MLX offline fallback
- **Git/Test Workflow Commands**: /test, /git, /pr for streamlined development
- **4 New Development Agents**: Rex, Dario, Otto, Paolo
- **Feature Release Manager Agent**: Automated quality gates
- **153 Compiler Warnings Fixed**: Production-quality zero-warning build
- **Budget/Theme Persistence**: Settings now persist correctly across sessions
- **Enhanced E2E Tests**: Coverage for new workflow commands and embeddings

## Technical Changes
- Fixed pthread_mutex_t type compatibility warnings
- Fixed sign conversion and double promotion warnings
- Updated README with v4.1.0 features and new REPL commands
- Fixed check_help_docs.sh to handle larger DETAILED_HELP arrays
- Updated VERSION to 4.1.0
- Updated CHANGELOG with comprehensive release notes

## Quality Gates Passed
- 51/51 fuzz tests
- 50/50 unit tests
- 39/39 compaction tests
- 32/32 documentation checks
- 0 compiler warnings

Developed by Roberto D'Angelo with the help of AI agents

+++ src/providers/mlx.m
@@ -608,7 +608,7 @@ MLXError mlx_generate(
 
     const MLXModelInfo* model = find_model_info(g_mlx_data.current_model_id);
     LOG_DEBUG(LOG_CAT_SYSTEM, "MLX: Generating response with %s (max_tokens=%zu, temp=%.2f)",
-              model ? model->display_name : "unknown", max_tokens, temperature);
+              model ? model->display_name : "unknown", max_tokens, (double)temperature);
 
     // Call Swift bridge for inference
     int32_t token_count = 0;
@@ -634,10 +634,10 @@ MLXError mlx_generate(
     }
 
     LOG_INFO(LOG_CAT_SYSTEM, "MLX: Generated %d tokens at %.1f tok/s",
-             token_count, tokens_per_sec);
+             token_count, (double)tokens_per_sec);
 
     // Update context tracking
-    g_mlx_data.context_used += mlx_estimate_tokens(prompt) + token_count;
+    g_mlx_data.context_used += mlx_estimate_tokens(prompt) + (size_t)token_count;
 
     *out_response = response;  // Caller must free()
